Build 3.9.11/20190902

+++ plugins/fabrik_element/emundus_fileupload/emundus_fileupload.js
@@ -155,6 +155,15 @@
                 if (xhr.readyState==4 && xhr.status==200) {
 
                     var result = JSON.parse(xhr.responseText);
+                    console.log(result.status);
+                    if(result.status == false){
+                        Swal.fire({
+                            type: 'error',
+                            title: Joomla.JText._('PLG_ELEMENT_FIELD_ERROR'),
+                            text: Joomla.JText._('PLG_ELEMENT_FIELD_ACCESS'),
+                            confirmButtonClass: 'btn btn-primary save-btn sauvegarder button save_continue'
+                        });
+                    }
                     for (var j = 0; j <= result.length; j++) {
 
                         if (result[j].ext == true && result[j].size == true && result[j].nbMax == true) {
@@ -165,8 +174,10 @@
                             div.appendChild(inputHidden);
 
                             Swal.fire({
-                                type: 'success',
+
                                 title: Joomla.JText._('PLG_ELEMENT_FIELD_SUCCESS'),
+                                text: Joomla.JText._('PLG_ELEMENT_FIELD_UPLOAD'),
+                                type: 'success',
                                 showConfirmButton: false,
                                 timer: 1500
                             });
@@ -175,8 +186,9 @@
                         if (result[j].ext == false) {
                             Swal.fire({
                                 type: 'error',
-                                title: 'Error',
-                                text: Joomla.JText._('PLG_ELEMENT_FIELD_EXTENSION')
+                                title: Joomla.JText._('PLG_ELEMENT_FIELD_ERROR'),
+                                text: Joomla.JText._('PLG_ELEMENT_FIELD_EXTENSION'),
+                                confirmButtonClass: 'btn btn-primary save-btn sauvegarder button save_continue'
                             });
 
                             input.value = '';
@@ -187,7 +199,8 @@
                             Swal.fire({
                                 type: 'error',
                                 title: Joomla.JText._('PLG_ELEMENT_FIELD_ERROR'),
-                                text: Joomla.JText._('PLG_ELEMENT_FIELD_ENCRYPT')
+                                text: Joomla.JText._('PLG_ELEMENT_FIELD_ENCRYPT'),
+                                confirmButtonClass: 'btn btn-primary save-btn sauvegarder button save_continue'
                             });
                             input.value = '';
                         }
@@ -197,6 +210,7 @@
                                 type: 'error',
                                 title: Joomla.JText._('PLG_ELEMENT_FIELD_ERROR'),
                                 text: Joomla.JText._('PLG_ELEMENT_FIELD_SIZE'),
+                                confirmButtonClass: 'btn btn-primary save-btn sauvegarder button save_continue'
                             });
                             input.value = '';
                             deleteButton.style.display = 'none';
@@ -206,7 +220,8 @@
                             Swal.fire({
                                 type: 'error',
                                 title: Joomla.JText._('PLG_ELEMENT_FIELD_ERROR'),
-                                text: Joomla.JText._('PLG_ELEMENT_FIELD_LIMIT')
+                                text: Joomla.JText._('PLG_ELEMENT_FIELD_LIMIT'),
+                                confirmButtonClass: 'btn btn-primary save-btn sauvegarder button save_continue'
                             });
                             input.value = '';
                             deleteButton.style.display = 'none';
@@ -236,32 +251,36 @@
 
                 if (xhr.readyState == 4 && xhr.status == 200) {
 
-                    var result = JSON.parse(xhr.responseText);
-                    for (var i = 0; i < result.length; i++) {
+                    if(xhr.responseText != '') {
 
-                        var divLink = document.createElement('div');
-                        divLink.setAttribute("id", elementId + '_attachment_link'+i);
-                        divLink.setAttribute("class", 'em-fileAttachment-link');
-                        divAttachment.appendChild(divLink);
+                        var result = JSON.parse(xhr.responseText);
 
-                        var link = document.createElement('a');
-                        var linkText = document.createTextNode(result[i].filename);
-                        link.setAttribute("href", result[i].target);
+                        for (var i = 0; i < result.length; i++) {
 
-                        divLink.appendChild(link);
-                        link.appendChild(linkText);
+                            var divLink = document.createElement('div');
+                            divLink.setAttribute("id", elementId + '_attachment_link' + i);
+                            divLink.setAttribute("class", 'em-fileAttachment-link');
+                            divAttachment.appendChild(divLink);
 
-                        var deleteButton = document.createElement('a');
-                        deleteButton.setAttribute("class", 'btn goback-btn em-deleteFile far fa-times-circle');
-                        deleteButton.setAttribute('value' , result[i].filename);
+                            var link = document.createElement('a');
+                            var linkText = document.createTextNode(result[i].filename);
+                            link.setAttribute("href", result[i].target);
 
-                        var icon = document.createElement('i');
-                        icon.setAttribute("class", 'far fa-times-circle');
+                            divLink.appendChild(link);
+                            link.appendChild(linkText);
 
-                        divLink.appendChild(deleteButton);
+                            var deleteButton = document.createElement('a');
+                            deleteButton.setAttribute("class", 'btn goback-btn em-deleteFile far fa-times-circle');
+                            deleteButton.setAttribute('value', result[i].filename);
 
-                        var button = document.querySelector('#'+elementId + '_attachment_link'+i+ ' > a.em-deleteFile');
-                        button.addEventListener('click', (event) => FbFileUpload.delete(elementId,attachId));
+                            var icon = document.createElement('i');
+                            icon.setAttribute("class", 'far fa-times-circle');
+
+                            divLink.appendChild(deleteButton);
+
+                            var button = document.querySelector('#' + elementId + '_attachment_link' + i + ' > a.em-deleteFile');
+                            button.addEventListener('click', (event) => FbFileUpload.delete(elementId, attachId));
+                        }
                     }
                 }
             };
@@ -291,7 +310,9 @@
                 confirmButtonColor: '#3085d6',
                 cancelButtonColor: '#d33',
                 confirmButtonText: Joomla.JText._('PLG_ELEMENT_FIELD_CONFIRM'),
-                cancelButtonText: Joomla.JText._('PLG_ELEMENT_FIELD_CANCEL')
+                confirmButtonClass: 'btn btn-primary save-btn sauvegarder button save_continue',
+                cancelButtonText: Joomla.JText._('PLG_ELEMENT_FIELD_CANCEL'),
+                cancelButtonClass: 'btn goback-btn button',
             }).then(answser => {
 
                 if (answser.value) {
@@ -302,11 +323,12 @@
                             var result = JSON.parse(xhr.responseText);
                             if (result.status == true) {
                                 parentDiv.remove();
-                                Swal.fire(
-                                    Joomla.JText._('PLG_ELEMENT_FIELD_DELETE'),
-                                    Joomla.JText._('PLG_ELEMENT_FIELD_DELETE_TEXT'),
-                                    'success'
-                                )
+                                Swal.fire({
+                                    title: Joomla.JText._('PLG_ELEMENT_FIELD_DELETE'),
+                                    text: Joomla.JText._('PLG_ELEMENT_FIELD_DELETE_TEXT'),
+                                    type: 'success',
+                                    confirmButtonClass: 'btn btn-primary save-btn sauvegarder button save_continue'
+                                });
                             }
                         }
                     };
