+++ plugins/fabrik_element/emundus_fileupload/emundus_fileupload.php
@@ -39,10 +39,10 @@ class PlgFabrik_ElementEmundus_fileupload extends PlgFabrik_Element {
     public $storage = null;
 
 
-	/**
-	 * @return bool
-	 * @throws Exception
-	 */
+    /**
+     * @return bool
+     * @throws Exception
+     */
     public function onBeforeStore() {
 
         jimport('joomla.filesystem.file');
@@ -51,7 +51,11 @@ class PlgFabrik_ElementEmundus_fileupload extends PlgFabrik_Element {
         $jinput = JFactory::getApplication()->input;
         $db = JFactory::getDBO();
 
-        if (EmundusHelperAccess::asApplicantAccessLevel($current_user->id)) {
+        if (JFactory::getUser()->guest) {
+
+            $status='false';
+            $result = array('status' => $status);
+            echo json_encode($result);
             return false;
         }
 
@@ -79,10 +83,13 @@ class PlgFabrik_ElementEmundus_fileupload extends PlgFabrik_Element {
         if (!empty($uploadResult)) {
             if ($nbMax == 1) {
                 $fileNameUpdate = $jinput->post->get($name.'_filename0');
-                $this->updateFile($current_user->fnum, $cid, $attachId, $fileNameUpdate);
+                if (!empty($fileNameUpdate)) {
+                    $this->updateFile($current_user->fnum, $cid, $attachId, $fileNameUpdate);
+                }
+
             }
             if ($nbMax > 1 && count($uploadResult) < $nbMax) {
-             	$this->insertFile($insert);
+                $this->insertFile($insert);
             }
         }
 
@@ -90,19 +97,23 @@ class PlgFabrik_ElementEmundus_fileupload extends PlgFabrik_Element {
     }
 
 
-	/**
-	 * @return bool
-	 */
+    /**
+     * @return bool
+     */
     public function onAjax_upload() {
+
         jimport('joomla.filesystem.file');
 
         $jinput = $this->app->input;
         $current_user = JFactory::getSession()->get('emundusUser');
 
-        if (EmundusHelperAccess::asApplicantAccessLevel($current_user->id)) {
-		    return false;
-	    }
-
+        if (JFactory::getUser()->guest) {
+            
+            $status='false';
+            $result = array('status' => $status);
+            echo json_encode($result);
+            return false;
+        }
         $attachId = $jinput->post->get('attachId');
         $can_submit_encrypted = $jinput->post->get('encrypt');
 
@@ -118,6 +129,21 @@ class PlgFabrik_ElementEmundus_fileupload extends PlgFabrik_Element {
         $nbMaxFile = (int)$attachmentResult->nbmax;
 
         $acceptedExt = [];
+
+        if (!file_exists(EMUNDUS_PATH_ABS.$current_user->id)) {
+            // An error would occur when the index.html file was missing, the 'Unable to create user file' error appeared yet the folder was created.
+            if (!file_exists(EMUNDUS_PATH_ABS.'index.html'))
+                touch(EMUNDUS_PATH_ABS.'index.html');
+
+            if (!mkdir(EMUNDUS_PATH_ABS.$current_user->id) || !copy(EMUNDUS_PATH_ABS.'index.html', EMUNDUS_PATH_ABS.$current_user->id.DS.'index.html')){
+                $error = JUri::getInstance().' :: USER ID : '.$current_user->id.' -> Unable to create user file';
+                JLog::add($error, JLog::ERROR, 'com_emundus');
+
+
+                return false;
+            }
+        }
+        chmod(EMUNDUS_PATH_ABS.$current_user->id, 0755);
 
         foreach ($files as $file) {
 
@@ -154,8 +180,8 @@ class PlgFabrik_ElementEmundus_fileupload extends PlgFabrik_Element {
                     if ($nbAttachment < $nbMaxFile) {
                         $nbMax = true;
                         if ($fileSize < $sizeMax) {
-	                        move_uploaded_file($tmp_name, $target);
-	                        $size = true;
+                            move_uploaded_file($tmp_name, $target);
+                            $size = true;
                         } else {
                             $size = false;
                         }
@@ -167,12 +193,13 @@ class PlgFabrik_ElementEmundus_fileupload extends PlgFabrik_Element {
                 if ($lengthFile > $nbMaxFile) {
                     $nbMax = false;
                 }
-
+                $result[] = array('size' => $size, 'ext' => $ext, 'nbMax' => $nbMax, 'filename' => $fileName, 'target' => $target,'nbAttachment' => $nbAttachment, 'encrypt' => $encrypt);
             } else {
                 $ext = false;
+                $result[] = array('size' => $size, 'ext' => $ext,  'filename' => $fileName, 'target' => $target,'nbAttachment' => $nbAttachment);
             }
 
-            $result[] = array('size' => $size, 'ext' => $ext, 'nbMax' => $nbMax, 'filename' => $fileName, 'target' => $target,'nbAttachment' => $nbAttachment, 'encrypt' => $encrypt);
+
         }
 
         echo json_encode($result);
@@ -180,17 +207,21 @@ class PlgFabrik_ElementEmundus_fileupload extends PlgFabrik_Element {
     }
 
 
-	/**
-	 * @return bool
-	 */
+    /**
+     * @return bool
+     */
     public function onAjax_attachment() {
 
         $jinput = $this->app->input;
 
         $current_user = JFactory::getSession()->get('emundusUser');
-	    if (EmundusHelperAccess::asApplicantAccessLevel($current_user->id)) {
-		    return false;
-	    }
+        if (JFactory::getUser()->guest) {
+
+            $status='false';
+            $result = array('status' => $status);
+            echo json_encode($result);
+            return false;
+        }
 
         $attachId = $jinput->post->get('attachId');
         $cid = $this->getCampaignId($current_user->fnum);
@@ -206,13 +237,14 @@ class PlgFabrik_ElementEmundus_fileupload extends PlgFabrik_Element {
         }
 
         echo json_encode($result);
-		return true;
+        return true;
+
     }
 
 
-	/**
-	 * @return bool
-	 */
+    /**
+     * @return bool
+     */
     public function onAjax_delete() {
 
         $jinput = $this->app->input;
@@ -221,21 +253,24 @@ class PlgFabrik_ElementEmundus_fileupload extends PlgFabrik_Element {
         $fileName = $jinput->post->get('filename');
         $attachId = $jinput->post->get('attachId');
 
-	    if (EmundusHelperAccess::asApplicantAccessLevel($current_user->id)) {
-		    return false;
-	    }
+        if (!EmundusHelperAccess::asApplicantAccessLevel($current_user->id) || !EmundusHelperAccess::asCoordinatorAccessLevel($current_user->id)) {
+            return false;
+        }
 
         $cid = $this->getCampaignId($current_user->fnum);
         $uploadResult = $this->getUploads($attachId, $current_user->id, $cid);
         $target = $this->getPath($current_user->id, $fileName);
 
-        if (file_exists($target)) {
+        if (file_exists($target) && !empty($uploadResult)) {
             unlink($target);
-            if (!empty($uploadResult)) {
-                $this->deleteFile($fileName, $current_user->fnum, $cid, $attachId);
-            }
+            $this->deleteFile($fileName, $current_user->fnum, $cid, $attachId);
             $status = true;
-        } else {
+        }
+        if (!file_exists($target) && !empty($uploadResult)) {
+            $this->deleteFile($fileName, $current_user->fnum, $cid, $attachId);
+            $status = true;
+        }
+        if (!file_exists($target) && empty($uploadResult)) {
             $status = false;
         }
 
@@ -244,41 +279,54 @@ class PlgFabrik_ElementEmundus_fileupload extends PlgFabrik_Element {
         return true;
     }
 
+    public function dataConsideredEmptyForValidation($data, $repeatCounter) {
+        $current_user = JFactory::getSession()->get('emundusUser');
 
-	/**
-	 * @return String
-	 * @throws Exception
-	 */
+        $attachId = $this->getAttachId();
+        $cid = $this->getCampaignId($current_user->fnum);
+        $uploadResult = $this->getUploads($attachId,$current_user->id,$cid);
+
+        if(empty($uploadResult) && $data == ""){
+            return true;
+        }
+        else{
+            return false;
+        }
+    }
+    /**
+     * @return String
+     * @throws Exception
+     */
     public function getFormId() {
         $jinput = JFactory::getApplication()->input;
         return $jinput->get('formid');
     }
 
 
-	/**
-	 * @return String
-	 * @throws Exception
-	 */
+    /**
+     * @return String
+     * @throws Exception
+     */
     public function getItemId() {
         $jinput = JFactory::getApplication()->input;
         return $jinput->get('Itemid', null);
     }
 
 
-	/**
-	 * @return mixed
-	 */
+    /**
+     * @return mixed
+     */
     public function getAttachId() {
         $params = $this->getParams();
         return $params->get('attachmentId');
     }
 
 
-	/**
-	 * @param $fnum
-	 *
-	 * @return Int
-	 */
+    /**
+     * @param $fnum
+     *
+     * @return Int
+     */
     public function getCampaignId($fnum) {
 
         $db = JFactory::getDBO();
@@ -293,11 +341,11 @@ class PlgFabrik_ElementEmundus_fileupload extends PlgFabrik_Element {
     }
 
 
-	/**
-	 * @param $attachId
-	 *
-	 * @return mixed
-	 */
+    /**
+     * @param $attachId
+     *
+     * @return mixed
+     */
     public function getAttachment($attachId) {
         $db = JFactory::getDBO();
 
@@ -311,45 +359,45 @@ class PlgFabrik_ElementEmundus_fileupload extends PlgFabrik_Element {
     }
 
 
-	/**
-	 * @param $attachId
-	 * @param $uid
-	 * @param $cid
-	 *
-	 * @return mixed
-	 */
+    /**
+     * @param $attachId
+     * @param $uid
+     * @param $cid
+     *
+     * @return mixed
+     */
     public function getUploads($attachId, $uid, $cid) {
         $db = JFactory::getDBO();
 
         $query = $db->getQuery(true);
         $query->select(array($db->quoteName('id'),$db->quoteName('filename')))
-	        ->from($db->quoteName('#__emundus_uploads'))
-	        ->where($db->quoteName('attachment_id') . ' = ' . $attachId . ' AND ' . $db->quoteName('user_id') . ' = ' . $uid . " AND " . $db->quoteName('campaign_id') . " = " . $cid);
+            ->from($db->quoteName('#__emundus_uploads'))
+            ->where($db->quoteName('attachment_id') . ' = ' . $attachId . ' AND ' . $db->quoteName('user_id') . ' = ' . $uid . " AND " . $db->quoteName('campaign_id') . " = " . $cid);
         $db->setQuery($query);
 
         return $db->loadObjectList();
     }
 
 
-	/**
-	 * @param $uid
-	 * @param $fileName
-	 *
-	 * @return string
-	 */
+    /**
+     * @param $uid
+     * @param $fileName
+     *
+     * @return string
+     */
     public function getPath($uid, $fileName) {
         return EMUNDUS_PATH_ABS.$uid.DS.$fileName;
     }
 
 
-	/**
-	 * @param $user
-	 * @param $attachId
-	 * @param $label
-	 * @param $file
-	 *
-	 * @return mixed
-	 */
+    /**
+     * @param $user
+     * @param $attachId
+     * @param $label
+     * @param $file
+     *
+     * @return mixed
+     */
     public function getFileName($user, $attachId, $label, $file) {
         $fileName = strtolower(preg_replace(array('([\40])', '([^a-zA-Z0-9-])', '(-{2,})'), array('_', '', '_'), preg_replace('/&([A-Za-z]{1,2})(grave|acute|circ|cedil|uml|lig);/', '$1', htmlentities(strtoupper($user->lastname) . '_' . ucfirst($user->firstname), ENT_NOQUOTES, 'UTF-8'))));
         $fileName .= '_' . $attachId . $label . '-' . rand() . '.' . pathinfo($file, PATHINFO_EXTENSION);
@@ -357,13 +405,13 @@ class PlgFabrik_ElementEmundus_fileupload extends PlgFabrik_Element {
     }
 
 
-	/**
-	 * @param             $data
-	 * @param   stdClass  $thisRow
-	 * @param   array     $opts
-	 *
-	 * @return string
-	 */
+    /**
+     * @param             $data
+     * @param   stdClass  $thisRow
+     * @param   array     $opts
+     *
+     * @return string
+     */
     public function ListData($data, stdClass &$thisRow, $opts = array()) {
         $profiler = JProfiler::getInstance('Application');
         JDEBUG ? $profiler->mark("renderListData: {$this->element->plugin}: start: {$this->element->name}") : null;
@@ -437,7 +485,7 @@ class PlgFabrik_ElementEmundus_fileupload extends PlgFabrik_Element {
      */
     public function render($data, $repeatCounter = 0) {
 
-    	JHTML::stylesheet('plugins/fabrik_element/emundus_fileupload/css/emundus_fileupload.css');
+        JHTML::stylesheet('plugins/fabrik_element/emundus_fileupload/css/emundus_fileupload.css');
         JHTML::script('plugins/fabrik_element/emundus_fileupload/emundus_fileupload.js');
 
         $params = $this->getParams();
@@ -644,6 +692,8 @@ class PlgFabrik_ElementEmundus_fileupload extends PlgFabrik_Element {
         JText::script('PLG_ELEMENT_FIELD_CANCEL');
         JText::script('PLG_ELEMENT_FIELD_DELETE');
         JText::script('PLG_ELEMENT_FIELD_DELETE_TEXT');
+        JText::script('PLG_ELEMENT_FIELD_ACCESS');
+        JText::script('PLG_ELEMENT_FIELD_UPLOAD');
 
         return array('FbField', $id, $opts);
     }
@@ -706,7 +756,7 @@ class PlgFabrik_ElementEmundus_fileupload extends PlgFabrik_Element {
         }
 
         switch ($p->get('text_format')) {
-	        default:
+            default:
             case 'text':
                 $objType = "VARCHAR(" . $p->get('maxlength', 255) . ")";
                 break;
@@ -963,37 +1013,42 @@ class PlgFabrik_ElementEmundus_fileupload extends PlgFabrik_Element {
     }
 
 
-	/**
-	 * @param $values
-	 *
-	 * @throws Exception
-	 */
-    public function insertFile($values) {
-        $db = JFactory::getDBO();
-        $query = $db->getQuery(true);
-        $columns = array('user_id', 'fnum', 'campaign_id', 'attachment_id', 'filename', 'can_be_deleted', 'can_be_viewed');
+    /**
+     * @param $values
+     *
+     * @throws Exception
+     */
+    public function insertFile($values)
+    {
+        if (!empty($values)) {
 
-        $query->insert($db->quoteName('#__emundus_uploads'))
-            ->columns($db->quoteName($columns))
-            ->values($values);
-        $db->setQuery($query);
 
-        try {
-            $db->execute();
-        } catch (Exception $e) {
-            JFactory::getApplication()->enqueueMessage('Probrème survenu au téléchargement des fichiers', 'message');
+            $db = JFactory::getDBO();
+            $query = $db->getQuery(true);
+            $columns = array('user_id', 'fnum', 'campaign_id', 'attachment_id', 'filename', 'can_be_deleted', 'can_be_viewed');
+
+            $query->insert($db->quoteName('#__emundus_uploads'))
+                ->columns($db->quoteName($columns))
+                ->values($values);
+            $db->setQuery($query);
+
+            try {
+                $db->execute();
+            } catch (Exception $e) {
+                JFactory::getApplication()->enqueueMessage('Probrème survenu au téléchargement des fichiers', 'message');
+            }
         }
     }
 
 
-	/**
-	 * @param $fnum
-	 * @param $cid
-	 * @param $attachId
-	 * @param $fileName
-	 *
-	 * @throws Exception
-	 */
+    /**
+     * @param $fnum
+     * @param $cid
+     * @param $attachId
+     * @param $fileName
+     *
+     * @throws Exception
+     */
     public function updateFile($fnum, $cid, $attachId, $fileName) {
         $db = JFactory::getDBO();
         $query = $db->getQuery(true);
@@ -1012,14 +1067,14 @@ class PlgFabrik_ElementEmundus_fileupload extends PlgFabrik_Element {
     }
 
 
-	/**
-	 * @param $fileName
-	 * @param $fnum
-	 * @param $cid
-	 * @param $attachId
-	 *
-	 * @throws Exception
-	 */
+    /**
+     * @param $fileName
+     * @param $fnum
+     * @param $cid
+     * @param $attachId
+     *
+     * @throws Exception
+     */
     public function deleteFile($fileName,$fnum, $cid, $attachId) {
         $db = JFactory::getDBO();
         $query = $db->getQuery(true);
@@ -1037,15 +1092,15 @@ class PlgFabrik_ElementEmundus_fileupload extends PlgFabrik_Element {
     }
 
 
-	/**
-	 * @param $file
-	 *
-	 * @return bool|false|int
-	 */
+    /**
+     * @param $file
+     *
+     * @return bool|false|int
+     */
     public function isEncrypted($file) {
         $f = fopen($file, 'rb');
         if (!$f) {
-	        return false;
+            return false;
         }
 
         //Read the last 320KB
